Move Model.Types.* classes in Model.Vocabulary package for better code consistency
