+++ netzob/src/netzob/Inference/Vocabulary/FormatOperations/ClusterByKeyField.py
@@ -47,11 +47,11 @@ from netzob.Common.Utils.Decorators import typeCheck, NetzobLogger
 from netzob.Model.Vocabulary.AbstractField import AbstractField
 from netzob.Model.Vocabulary.Field import Field
 from netzob.Model.Vocabulary.Symbol import Symbol
-from netzob.Model.Types.TypeConverter import TypeConverter
-from netzob.Model.Types.HexaString import HexaString
-from netzob.Model.Types.ASCII import ASCII
-from netzob.Model.Types.BitArray import BitArray
-from netzob.Model.Types.Raw import Raw
+from netzob.Model.Vocabulary.Types.TypeConverter import TypeConverter
+from netzob.Model.Vocabulary.Types.HexaString import HexaString
+from netzob.Model.Vocabulary.Types.ASCII import ASCII
+from netzob.Model.Vocabulary.Types.BitArray import BitArray
+from netzob.Model.Vocabulary.Types.Raw import Raw
 from netzob.Common.Utils.DataAlignment.DataAlignment import DataAlignment
