+++ netzob/src/netzob/Inference/Vocabulary/FormatOperations/ClusterBySize.py
@@ -47,9 +47,9 @@ from netzob.Common.Utils.Decorators import typeCheck, NetzobLogger
 from netzob.Model.Vocabulary.AbstractField import AbstractField
 from netzob.Model.Vocabulary.Field import Field
 from netzob.Model.Vocabulary.Symbol import Symbol
-from netzob.Model.Types.TypeConverter import TypeConverter
-from netzob.Model.Types.HexaString import HexaString
-from netzob.Model.Types.Raw import Raw
+from netzob.Model.Vocabulary.Types.TypeConverter import TypeConverter
+from netzob.Model.Vocabulary.Types.HexaString import HexaString
+from netzob.Model.Vocabulary.Types.Raw import Raw
 from netzob.Model.Vocabulary.Messages.RawMessage import RawMessage
 from netzob.Model.Vocabulary.Domain.DomainFactory import DomainFactory
