+++ netzob/src/netzob/Inference/Vocabulary/FormatOperations/FieldSplitAligned/FieldSplitAligned.py
@@ -43,10 +43,10 @@ from collections import OrderedDict
 from netzob.Common.Utils.Decorators import typeCheck, NetzobLogger
 from netzob.Model.Vocabulary.AbstractField import AbstractField
 from netzob.Common.C_Extensions.WrapperArgsFactory import WrapperArgsFactory
-from netzob.Model.Types.AbstractType import AbstractType
-from netzob.Model.Types.TypeConverter import TypeConverter
-from netzob.Model.Types.HexaString import HexaString
-from netzob.Model.Types.Raw import Raw
+from netzob.Model.Vocabulary.Types.AbstractType import AbstractType
+from netzob.Model.Vocabulary.Types.TypeConverter import TypeConverter
+from netzob.Model.Vocabulary.Types.HexaString import HexaString
+from netzob.Model.Vocabulary.Types.Raw import Raw
 from netzob.Model.Vocabulary.Field import Field
 from netzob.Model.Vocabulary.Messages.AbstractMessage import AbstractMessage
 from netzob.Inference.Vocabulary.Search.SearchEngine import SearchEngine
