+++ netzob/src/netzob/Inference/Vocabulary/FormatOperations/FieldSplitStatic/ParallelFieldSplitStatic.py
@@ -40,7 +40,7 @@ import multiprocessing
 #| Local application imports
 #+---------------------------------------------------------------------------+
 from netzob.Common.Utils.Decorators import typeCheck, NetzobLogger
-from netzob.Model.Types.AbstractType import AbstractType
+from netzob.Model.Vocabulary.Types.AbstractType import AbstractType
 from netzob.Model.Vocabulary.AbstractField import AbstractField
 from netzob.Common.Utils.DataAlignment.DataAlignment import DataAlignment
