+++ netzob/src/netzob/Model/Vocabulary/Domain/Parser/FlowParser.py
@@ -47,9 +47,9 @@ from netzob.Model.Vocabulary.Messages.RawMessage import RawMessage
 
 from netzob.Model.Vocabulary.Symbol import Symbol
 from netzob.Model.Vocabulary.Domain.Parser.ParsingPath import ParsingPath
-from netzob.Model.Types.TypeConverter import TypeConverter
-from netzob.Model.Types.BitArray import BitArray
-from netzob.Model.Types.Raw import Raw
+from netzob.Model.Vocabulary.Types.TypeConverter import TypeConverter
+from netzob.Model.Vocabulary.Types.BitArray import BitArray
+from netzob.Model.Vocabulary.Types.Raw import Raw
 from netzob.Model.Vocabulary.Domain.Parser.FieldParser import FieldParser
