+++ netzob/src/netzob/Model/Vocabulary/Types/Timestamp.py
@@ -45,10 +45,10 @@ from bitarray import bitarray
 # | Local application imports                                                 |
 # +---------------------------------------------------------------------------+
 from netzob.Common.Utils.Decorators import typeCheck, NetzobLogger
-from netzob.Model.Types.AbstractType import AbstractType
-from netzob.Model.Types.TypeConverter import TypeConverter
-from netzob.Model.Types.BitArray import BitArray
-from netzob.Model.Types.Integer import Integer
+from netzob.Model.Vocabulary.Types.AbstractType import AbstractType
+from netzob.Model.Vocabulary.Types.TypeConverter import TypeConverter
+from netzob.Model.Vocabulary.Types.BitArray import BitArray
+from netzob.Model.Vocabulary.Types.Integer import Integer
 
 
 @NetzobLogger
@@ -283,7 +283,7 @@ class Timestamp(AbstractType):
                unitSize=AbstractType.UNITSIZE_32,
                endianness=AbstractType.defaultEndianness(),
                sign=AbstractType.SIGN_UNSIGNED):
-        from netzob.Model.Types.Raw import Raw
+        from netzob.Model.Vocabulary.Types.Raw import Raw
 
         intValue = TypeConverter.convert(
             data,
