+++ netzob/src/netzob/Model/all.py
@@ -31,7 +31,7 @@
 
 from netzob.Model.Vocabulary.all import *
 from netzob.Model.Grammar.all import *
-from netzob.Model.Types.all import *
+from netzob.Model.Vocabulary.Types.all import *
 from netzob.Simulator.all import *
 
 from netzob.Model.Protocol import Protocol
